+++ benchexec/__init__.py
@@ -49,7 +49,7 @@ Variables ending with "tag" contain references to XML tag objects created by the
 # prepare for Python 3
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-__version__ = '1.11-dev'
+__version__ = '1.13-dev'
 
 class BenchExecException(Exception):
     pass