+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # BenchExec Changelog
 
+## BenchExec 1.12
+
+- Fix execution of runs specified with `<withoutfile>` tags
+  in the benchmark definition: the name of the run was missing
+  from the command-line in BenchExec 1.11.
+
+## BenchExec 1.11
+
+- `table-generator` can now be given result XML files as arguments
+  in addition to a table-definition XML file (with parameter `-x`).
+  In this case, it will use the column definitions from the latter
+  for tables with the separately given results.
+- The directory `contrib` of the repository now contains a script
+  [`statistics-tex.py`](https://github.com/sosy-lab/benchexec/blob/master/contrib/statistics-tex.py),
+  which can export summary data for benchmark results
+  (e.g., number of solved tasks, average CPU time, etc.)
+  to LaTeX.
+- The dummy tools `true` and `false`, which could be used for testing
+  a BenchExec installation, are replaced with a more generic dummy tool
+  called `dummy`.
+- A few minor bug fixes and performance optimizations.
+
+A new paper about BenchExec called
+[Reliable Benchmarking: Requirements and Solutions](https://www.sosy-lab.org/~dbeyer/Publications/2017-STTT.Reliable_Benchmarking_Requirements_and_Solutions.pdf)
+is now available.
+
+Please note that support for Python 3.2 and 3.3 is deprecated.
+Furthermore, the support for "sudo mode" (parameter `--user`/`--users`)
+is also deprecated.
+All deprecated features will be removed in BenchExec 2.0.
+
+
 ## BenchExec 1.10
 
 This release brings several smaller and medium-sized features:
@@ -19,12 +51,12 @@ This release brings several smaller and medium-sized features:
 - The feature for executing benchmarks under different user accounts with sudo
   is now marked as deprecated and may be removed in the future,
   consider using the container mode instead for isolating runs
-  (cf. [issue #215](https://github.com/sosy-lab/benchexec/issues/215).
+  (cf. [issue #215](https://github.com/sosy-lab/benchexec/issues/215)).
 - `table-generator` is now more flexible:
   - Builtin support for certain unit conversions,
     such that the scale factor does not always need to be explicitly specified.
     Furthermore, unit conversions now work even if the values already have a unit.
-  - Column titles can be manually specified with the `displayTitle` attribute`.
+  - Column titles can be manually specified with the `displayTitle` attribute.
   - What columns are relevant for the "diff" table can be configured.
 
 Please also note that we are considering dropping the support for Python 3.2